// src/models/tmdb/movie/MovieSearchElement.ts
import { MediaType } from "../enum";

export interface MovieSearchElement {
  adult: boolean;
  backdrop_path: string;
  id: number;
  title: string;
  original_language: string;
  original_title: string;
  overview: string;
  poster_path: string;
  genre_ids: Array<number>;
  popularity: number;
  release_date: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  media_type: MediaType;
}